Anchor right when overlapping the hovered element
The :hover styling wasn't working for me in practice, so I opted to give
this route a try. Probably bears playing with for a while. I'm also
ignoring pointer events on the statusbar, so if all else fails and we
end up with the bar over the hovered element, it's still possible to
click the element through the bar.

(I don't remove the right-side class in hideStatus(), since then it could
jump around during the fade-out transition.)